examples/sites/syntaxviewer: initial commit
Syntax viewer is an incredibly simple application for viewing the syntax
tree representations of different languages.

As part of this we vendor mvdan.cc/sh/syntax and its deps.

We also re-enable AriaSet for now until we get
https://github.com/gopherjs/gopherjs/pull/779 merged.

We also add the float CSS attribute.

And fix some of the JSX parsing.